chore: refactor into hooks in features/browser
