+++ packages/extension/src/ui/App.tsx
@@ -1,12 +1,12 @@
 import { ThemeProvider, createTheme } from "@mui/material"
-import { FC, Suspense, useEffect, useState } from "react"
+import { FC, Suspense } from "react"
 import { createGlobalStyle } from "styled-components"
 import { normalize } from "styled-normalize"
 import { SWRConfig } from "swr"
 
 import { AppRoutes } from "./AppRoutes"
 import { LoadingScreen } from "./features/actions/LoadingScreen"
-import { isInTab } from "./features/recovery/useCustomNavigate"
+import { useExtensionIsInTab } from "./features/browser/tabs"
 import { swrCacheProvider } from "./services/swr"
 
 const GlobalStyleWithFixedDimensions = createGlobalStyle`
@@ -54,14 +54,11 @@ const OverwriteDimensionsToMinDimensions = createGlobalStyle`
 `
 
 const GlobalStyle: FC = () => {
-  const [isTab, setIsTab] = useState(false)
-  useEffect(() => {
-    isInTab().then(setIsTab)
-  }, [])
+  const extensionIsInTab = useExtensionIsInTab()
   return (
     <>
       <GlobalStyleWithFixedDimensions />
-      {isTab && <OverwriteDimensionsToMinDimensions />}
+      {extensionIsInTab && <OverwriteDimensionsToMinDimensions />}
     </>
   )
 }
